Fix minor bug and clarify string length behavior.
Fixes a bug where the two-step string example included a null
terminator.

I'm a bit unsure of line 127, and if it's valid to set length
to the exact end of the string, or if that's invalid, because
that memory does contain a null terminator.

However, this should do a better job of expalining how the two-step
string allocation process works.